backend/requirements.txt
# ACAD=GIS API Server Requirements
# Install with: pip install -r requirements.txt

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
python-dotenv==1.0.0

# Data validation
pydantic==2.5.0

# CORS middleware (included in fastapi, but listed for clarity)
# Optional: Add these if you need them later
# openai==1.3.0  # For embeddings generation
# numpy==1.26.2  # For vector operations
